use premake for make generator

+++ core/include/StringUtil.h
@@ -12,8 +12,8 @@ class StringUtil{
 		static bool endsWith(const string str, const string endsWithStr);
 		static string replaceAll(string& str, const string& oldstr, const string& newstr);
 		static string escape(string& str);
-		static vector<string> split(const string& str, const string& delimiters, const int maxParts=0);
-		static vector<string> split(const string& str, const string& delimiters, const int maxParts, vector<string>& tokens);
+		static vector<string> split(const string& str, const string& delimiters, const size_t maxParts=0);
+		static vector<string> split(const string& str, const string& delimiters, const size_t maxParts, vector<string>& tokens);
 		static string trim(string& str, string toTrimString=" ");
 		static string trimLeft(string& str, string toTrimString=" \t", bool trimAll = true);
 		static string trimRight(string& str, string toTrimString=" \t", bool trimAll = true);
